[INJICERT-434] fix TemplateRepository definition
Co-Authored by: Hitesh Jain <jainhitesh9998@gmail.com>
