+++ certify-service/src/main/resources/application-local.properties
@@ -203,7 +203,6 @@ mosip.certify.mock.authenticator.get-identity-url="http://localhost:8082/v1/mock
 mosip.certify.mock.vciplugin.issuer.key-cert="dummy-issuer-cert"
 mosip.certify.mock.vciplugin.ca.key-cert="dummy-ca-cert"
 
-
 ## ---------------------------------------- Cache configuration --------------------------------------------------------
 
 mosip.certify.cache.secure.individual-id=true
